src/sage/geometry/cone.py: new irreducible_factors() method for cones
This was effectively suggested in PR 40367 to compute linear
isomorphisms of cones. To find isomorphisms, we would like to
decompose the cones into irreducible factors, and then look for
isomorphisms between factors of the appropriate size. Since the
isomorphism algorithm is combinatorial, it is better to work with lots
of small cones than it is to work with one big cone.

In any case, this is of independent interest and is not too hard to
implement. Thanks to Dmitrii Pasechnik for the suggestion (and the
algorithm).